requirements.txt
flask
openai
nltk
PyPDF2
python-dotenv
